feat: Update company and professional models
This commit updates the company and professional models. It renames the `company.ts` file to `shared/models/company.ts` and moves the `professional.ts` file to `shared/models/professional_type.ts`. It also creates a new file `shared/models/professional.ts` with the necessary model definitions. These changes are made to improve the organization and structure of the codebase.

Note: The recent user commits and repository commits are not relevant for generating the commit message.
