+++ apps/shared/models/professional_type.ts
@@ -2,24 +2,12 @@ import { BaseModel, beforeCreate, column } from '@adonisjs/lucid/orm'
 import { DateTime } from 'luxon'
 import { generateSnowflake } from '#apps/shared/services/snowflake_service'
 
-export default class Professional extends BaseModel {
+export default class ProfessionalType extends BaseModel {
   @column({ isPrimary: true })
   declare id: string
 
   @column()
-  declare oidcId: string
-
-  @column()
-  declare type: string // ProfessionalType
-
-  @column()
-  declare identityId: string
-
-  @column()
-  declare contactId: string
-
-  @column()
-  declare addressId: string
+  declare name: string
 
   @column.dateTime({ autoCreate: true })
   declare createdAt: DateTime
@@ -28,7 +16,7 @@ export default class Professional extends BaseModel {
   declare updatedAt: DateTime | null
 
   @beforeCreate()
-  static async generateUuid(model: Professional) {
+  static async generateUuid(model: ProfessionalType) {
     model.id = generateSnowflake()
   }
 }
